+++ friendlysam/tests/test_state_vars.py
@@ -22,7 +22,7 @@ def test_state_vars():
     cl = Cluster(p, c, resource=RESOURCE, name='Cluster')
 
     prob = fs.Problem()
-    prob += (part.constraints(t) for part, t in product(cl.descendants_and_self, times))
+    prob += (part.constraints.make(t) for part, t in product(cl.descendants_and_self, times))
 
     prob.objective = fs.Minimize(sum(p.cost(t) for t in times))