+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [unreleased] - unreleased
 
+=======
+### Fixed
+- Fixed a serious bug in `FlowNetwork`/`Node` implementation causing problems if one `Node` was connected in several `FlowNetwork`s.
+
+
 ## [0.2.0] - 2015-06-08
 
 ### Added