+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Added
  - Added `ub`, `lb`, and `domain` to `VariableCollection`.
+ - Added `long_description` to `Constraint`.
 
 ### Fixed
 - Fixed a serious bug in `FlowNetwork`/`Node` implementation causing problems if one `Node` was connected in several `FlowNetwork`s.