+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Changed
  - Moved some `__init__` code to `__new__` to simplify inheritance of e.g. `Node`.
+ - Node.resources now contains keys from inflows and outflows.
 
 ### Fixed
 - Fixed a serious bug in `FlowNetwork`/`Node` implementation causing problems if one `Node` was connected in several `FlowNetwork`s.